requirements.txt
functions-framework==3.*
psycopg2-binary==2.9.9
requests